Run Magisk patches in an aarch64 virtual machine
This commit adds a script to build a virtual machine image that is used
by avbroot for running the Magisk patches. This allows magiskboot to run
in a controlled environment that is much more similar to what it
encounters on a real device. This also makes avbroot truly cross
platform.

The virtual machine image is based on Alpine Linux (both the kernel and
userspace). magiskboot currently has no dependency on an Android kernel
or userspace. Android's prebuilt GKI kernel unfortunately is not
compiled with support for devtmpfs and 9p, so if magiskboot ever starts
requiring Android kernel features, we'll have to build a custom kernel.
As for userspace, the only soft dependency is for getprop and that is
only executed if /system/build.prop does not exist. The VM image
includes a minimal build.prop to satisfy that requirement.

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>