Strip out 0xd935 extra field record in zip entries
The extra field record was previously stripped out due to a bug in
Python's zipfile implementation. That was fixed by [1], which was
backported to 3.10 and 3.11, but not 3.9.

This commit explicitly strips out the 0xd935 record to ensure that the
checksums of the patched files are the same across all Python versions.
This is probably the correct thing to do anyway, since the record is
related to zip entry alignment [2], which we don't do.

Note that the Python bug can affect other extra field records too, but
we don't see that in practice.

[1] https://github.com/python/cpython/commit/62c0327487ce39ec6504919b7712e0f197075ef9
[2] https://android.googlesource.com/platform/build/+/android-13.0.0_r1/tools/signapk/src/com/android/signapk/SignApk.java#133

Fixes: #93

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>