cpio: Read data by incremental reallocation after a threshold
This prevents avbroot from being killed for excessive memory usage due
to the entry header specifying a very large file name or file size
(larger than what's actually contained in the file).

This threshold for switching to incremental reallocation is 1 KiB for
file names and 1 MiB for file data.

Issue: #157

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>