[BROKEN] Proof of concept implementation of option 5
This adds a wrapper for /init that spawns a child process to ptrace the
parent. When stage 1 execs stage 2, the tracer will bind mount the new
otacerts.zip, detach, and exit.

The mount process works, but stage 2 panics and reboots to the
bootloader for unknown reasons. The conditions that lead to the reboot
don't result in the kernel log being preserved.

Issue: #225

Signed-off-by: Andrew Gunnerson <accounts+github@chiller3.com>